Fixed bugs with assets
* fixed bug which could not link to the assets folder.

* now the module trigger phrase is dynamic, the extension will read from
  the config file and extract the MODULE_ASSET_TRIGGER and dynamically
  insert into the newly created module.

* added parent::__constructor(); that can allow the super module use its
  parrent class such as model.

* adjusted the template.
