+++ src/commands/templates/trongate-view.template.ts
@@ -1,4 +1,6 @@
-export function viewTemplate(moduleName: string) {
+export function viewTemplate(moduleName: string, GLOBAL_SETTINGS: any) {
+    const assetsTriggerPhrase = GLOBAL_SETTINGS['config']['MODULE_ASSETS_TRIGGER']
+    
     return `<h1>Hello from <span><i>${moduleName}</i></span> module</h1>
     <p>This view was generated using Trongate Scaffold & Code Snippets!</p>
         <div>
@@ -11,7 +13,7 @@ export function viewTemplate(moduleName: string) {
         </div>
 <style>
     /* Link to the ${moduleName}/assets/css/custom.css */
-    @import url('<?= BASE_URL ?>${moduleName}/css/custom.css');
+    @import url('<?= BASE_URL ?>${GLOBAL_SETTINGS['superModule']?`${GLOBAL_SETTINGS['parentModuleName']}-${moduleName}${assetsTriggerPhrase}`:`${moduleName}${assetsTriggerPhrase}`}/css/custom.css');
 
     body {
         background-color: rgb(251, 178, 165);
@@ -19,7 +21,7 @@ export function viewTemplate(moduleName: string) {
 </style>
 
 <!-- Link to the assets/js/custom.js -->
-<script src="<?= BASE_URL ?>${moduleName}/js/custom.js"></script>`;
+<script src="<?= BASE_URL ?>${GLOBAL_SETTINGS['superModule']?`${GLOBAL_SETTINGS['parentModuleName']}-${moduleName}${assetsTriggerPhrase}`:`${moduleName}${assetsTriggerPhrase}`}/js/custom.js"></script>`;
 }
 
 export function getTrongateModuleCss() {
